Revamp the ed25519 API
This is a breaking API change, to make the ed25519 API similiar to
the one used for ECDSA, and for future signature systems.

The ed25519 API was difficult to extend. In particular, adding
multiple encodings for keys would have been ugly.

Instead of raw bytes, keys are now represented as PublicKey and
SecretKey types, and signatures as the Signature type.

This is consistent to the ECDSA API, and, overall, makes everything
much cleaner.

Non-batch signature creation and verification also got a little
bit faster with the removal of an unneeded serialization.

Blind key signatures now directly reuse the regular signature code
instead of duplicating it just to use a different way to compute
the prefix.

In addition, support for incremental signatures have been added.
Very large messages can now be signed and verified without
requiring much memory.

This is something that was requested before, but was blocked by
previous API.