Sema: tweaked strategy for generic instantiation failure
This is a similar strategy as @Vexu's recent solution to the same
problem, but tweaked in the following ways:

 * It is a bit less invasive in the sense that it does not extract a
   large chunk of code into a new function.
 * It does mostly the same removal of `errdefer`, however it leaves in
   the removal of the decl from the `monomorphed_funcs` table in the
   case of an error.
 * The changes to Decl memory handling are now done the same in Sema as
   they are in Module.

Some of the memory allocation failures in Sema are not handled properly.
Instead of pretending with `try`, the code now acknowledges this fact by
using `@panic`.

I have a long term plan for how to fix these, which is to migrate
everything over to InternPool. Then we can do the "ensure capacity"
pattern for all memory allocations needed by generic instantiation.

I'm not sure why the compile error doesn't match exactly in the
generic_instantiation_failure_container_level test case, but in my
opinion this logic should be in AstGen anyway, not Sema.